OrderHistoryDetail and Review Complete

+++ app/src/main/java/com/os/foodie/data/network/model/order/restaurant/detail/OrderHistoryDetail.java
@@ -51,7 +51,7 @@ public class OrderHistoryDetail {
         private String deliveryType;
         @SerializedName("discount")
         @Expose
-        private Integer discount;
+        private Float discount;
         @SerializedName("is_reviewed")
         @Expose
         private String isReviewed;
@@ -120,11 +120,11 @@ public class OrderHistoryDetail {
             this.deliveryType = deliveryType;
         }
 
-        public Integer getDiscount() {
+        public Float getDiscount() {
             return discount;
         }
 
-        public void setDiscount(Integer discount) {
+        public void setDiscount(Float discount) {
             this.discount = discount;
         }
